+++ terraform/aws/analytical-platform-data-engineering-production/10ds/vpc.tf
@@ -1,6 +1,6 @@
 module "vpc" {
   source  = "terraform-aws-modules/vpc/aws"
-  version = "5.1.1"
+  version = "5.1.2"
 
   name                   = local.name
   cidr                   = local.vpc_cidr