+++ terraform/aws/analytical-platform/oidc/oidc-roles.tf
@@ -86,7 +86,7 @@ data "aws_iam_policy_document" "github_oidc_role" {
 module "analytical_platform_data_engineering_production_github_oidc_role" {
   for_each = { for role, config in local.oidc_roles : role => config if try(config.account, "analytical-platform-management-production") == "analytical-platform-data-engineering-production" }
 
-  source = "github.com/ministryofjustice/modernisation-platform-github-oidc-role?ref=v2.0.1"
+  source = "github.com/ministryofjustice/modernisation-platform-github-oidc-role?ref=v3.0.0"
 
   providers = {
     aws = aws.analytical-platform-data-engineering-production
@@ -102,7 +102,7 @@ module "analytical_platform_data_engineering_production_github_oidc_role" {
 module "analytical_platform_data_engineering_sandbox_a_github_oidc_role" {
   for_each = { for role, config in local.oidc_roles : role => config if try(config.account, "analytical-platform-management-production") == "analytical-platform-data-engineering-sandbox-a" }
 
-  source = "github.com/ministryofjustice/modernisation-platform-github-oidc-role?ref=v2.0.0"
+  source = "github.com/ministryofjustice/modernisation-platform-github-oidc-role?ref=v3.0.0"
 
   providers = {
     aws = aws.analytical-platform-data-engineering-sandbox-a
@@ -134,7 +134,7 @@ module "analytical_platform_data_production_github_oidc_role" {
 module "analytical_platform_management_production_github_oidc_role" {
   for_each = { for role, config in local.oidc_roles : role => config if try(config.account, "analytical-platform-management-production") == "analytical-platform-management-production" }
 
-  source = "github.com/ministryofjustice/modernisation-platform-github-oidc-role?ref=v2.0.0"
+  source = "github.com/ministryofjustice/modernisation-platform-github-oidc-role?ref=v3.0.0"
 
   providers = {
     aws = aws.analytical-platform-management-production