+++ terraform/pagerduty/terraform.tf
@@ -10,11 +10,11 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = "5.15.0"
+      version = "5.17.0"
     }
     pagerduty = {
       source  = "pagerduty/pagerduty"
-      version = "2.16.2"
+      version = "3.0.0"
     }
   }
 }